refactor the look
