--- frontend/src/editor/style.css
@@ -1,3 +1,0 @@
-.form-group:not(:last-child) {
-    border-bottom: 1px solid rgba(0, 0, 0, 0.1);
-}
